Moves Configurator class to Configurator.js
